requirements.txt
aioredis==2.0.1
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
databases==0.8.0
exceptiongroup==1.1.3
fastapi==0.103.2
greenlet==3.0.0
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
Pillow==10.0.1
psycopg2==2.9.9
pycparser==2.21
pydantic==2.4.2
pydantic-core==2.10.1
PyJWT==2.8.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
typing-extensions==4.8.0
uvicorn==0.23.2
